[REFACTOR/#61] 알람 소리 관련 기능을 core:media 모듈로 이전
