+++ core/media/build.gradle.kts
@@ -1,0 +1,11 @@
+import com.yapp.convention.setNamespace
+
+plugins {
+    id("orbit.android.library")
+    id("orbit.android.hilt")
+    id("kotlinx-serialization")
+}
+
+android {
+    setNamespace("core.media")
+}
