+++ data/src/main/java/com/yapp/data/local/repositoryimpl/AlarmRepositoryImpl.kt
@@ -1,41 +1,22 @@
 package com.yapp.data.local.repositoryimpl
 
-import android.content.Context
 import com.yapp.data.local.datasource.AlarmLocalDataSource
-import com.yapp.data.local.datasource.SoundPlayer
 import com.yapp.data.local.toEntity
 import com.yapp.domain.model.Alarm
 import com.yapp.domain.model.AlarmSound
 import com.yapp.domain.repository.AlarmRepository
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
+import com.yapp.media.ringtone.RingtoneManagerHelper
+import com.yapp.media.sound.SoundPlayer
 import javax.inject.Inject
 
 class AlarmRepositoryImpl @Inject constructor(
     private val alarmLocalDataSource: AlarmLocalDataSource,
+    private val ringtoneManagerHelper: RingtoneManagerHelper,
     private val soundPlayer: SoundPlayer,
-    @ApplicationContext private val context: Context,
 ) : AlarmRepository {
-    override suspend fun getAlarmSounds(): Result<List<AlarmSound>> = withContext(Dispatchers.IO) {
-        runCatching {
-            val ringtoneManager = android.media.RingtoneManager(context).apply {
-                setType(android.media.RingtoneManager.TYPE_ALARM)
-            }
-
-            val cursor = ringtoneManager.cursor
-            val sounds = mutableListOf<AlarmSound>()
-
-            cursor.use {
-                if (it.moveToFirst()) {
-                    do {
-                        val title = cursor.getString(android.media.RingtoneManager.TITLE_COLUMN_INDEX)
-                        val uri = ringtoneManager.getRingtoneUri(it.position)
-                        sounds.add(AlarmSound(title, uri))
-                    } while (it.moveToNext())
-                }
-            }
-            sounds
+    override suspend fun getAlarmSounds(): Result<List<AlarmSound>> = runCatching {
+        ringtoneManagerHelper.getAlarmSounds().map { (title, uri) ->
+            AlarmSound(title, uri)
         }
     }
