+++ settings.gradle.kts
@@ -40,3 +40,4 @@ include(":feature:navigator")
 include(":feature:mypage")
 include(":feature:onboarding")
 include(":feature:mission")
+include(":core:media")
